+++ CHANGES.md
@@ -3,6 +3,7 @@
 - Combined all the provider traits(Store/Expiry/ExpiryStore) into a single trait named Provider
 - Removed the glue code and no_expiry method
 - Renamed `store` method to `provider` in builder
+- Changed `mutate` to return the value
 
 ## Version 0.4 Alpha.4
 - Fixed set_expiring method signature